refactor: adjust customers domain & mv keys inside hook
